# order-service/docs/swagger.yaml
basePath: /
definitions:
...
  models.Order:
    properties:
      confirmed_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      inventory_confirmed:
        type: boolean
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      payment_confirmed:
        type: boolean
      payment_id:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.OrderItem:
    properties:
      id:
        type: string
      order_id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      subtotal:
        type: number
      unit_price:
        type: number
    type: object
host: localhost:8083
info:
  contact: {}
  description: Core service responsible for managing order life cycles and saga coordination.
  title: Order Service API
  version: "1.0"
paths:
...
  /user/{userId}:
    get:
      description: Retrieve a list of all orders associated with a specific User UUID.
      parameters:
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "400":
          description: invalid user_id format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: failed to fetch user orders
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get orders by user ID
      tags:
      - orders
swagger: "2.0"